Test data necessary for python samples added to test package. Test launch script updated to run new tests as well.
